don't update Bundler's cache symlink if it exists and is pointing to the right place
I am hitting issues with multiple `autoproj exec` instances
being started in parallel, with some of them fail in this
codepath because they all try to re-create the symlink at the
same time.

This is a poor man's fix for it, basically assuming that the
symlink, once created, won't change.